Push color column to Table component
